chore: dont delete correct cli
