--- apps/cli-old/.eslintignore
@@ -1,2 +1,0 @@
-/lib
-/test
