--- apps/cli-old/.gitignore
@@ -1,8 +1,0 @@
-*-debug.log
-*-error.log
-/.nyc_output
-/dist
-/lib
-/tmp
-/yarn.lock
-node_modules
