--- apps/cli-old/package.json
@@ -1,82 +1,0 @@
-{
-  "name": "@inlang/cli-old",
-  "description": "TODO",
-  "version": "0.1.0",
-  "author": "inlang",
-  "bin": {
-    "inlang": "./bin/run"
-  },
-  "bugs": "https://github.com/inlang/inlang/issues",
-  "dependencies": {
-    "@oclif/command": "^1.8.6",
-    "@oclif/config": "^1.18.2",
-    "@oclif/plugin-help": "^5.1.10",
-    "@oclif/plugin-not-found": "^2.2.4",
-    "@types/cli-color": "^2.0.1",
-    "@types/node-fetch": "^2.5.12",
-    "cli-color": "^2.0.1",
-    "glob": "^7.2.0",
-    "iso-639-2": "^3.0.1",
-    "lodash": "^4.17.21",
-    "node-fetch": "^3.1.1",
-    "prompts": "^2.4.2",
-    "tslib": "^1.14.1"
-  },
-  "devDependencies": {
-    "@inlang/adapters": "*",
-    "@inlang/common": "*",
-    "@inlang/database": "*",
-    "@inlang/eslint-config": "*",
-    "@inlang/fluent-syntax": "*",
-    "@oclif/dev-cli": "^1.26.3",
-    "@types/lodash": "^4.14.176",
-    "@types/node": "^10.17.60",
-    "@typescript-eslint/eslint-plugin": "^5.4.0",
-    "@typescript-eslint/parser": "^5.4.0",
-    "ts-node": "^8.10.2",
-    "webpack": "^5.64.4",
-    "webpack-cli": "^4.9.1"
-  },
-  "engines": {
-    "node": ">=8.0.0"
-  },
-  "files": [
-    "/bin",
-    "/lib",
-    "/npm-shrinkwrap.json",
-    "/oclif.manifest.json"
-  ],
-  "homepage": "https://github.com/inlang/inlang",
-  "keywords": [
-    "inlang",
-    "cli",
-    "i18n",
-    "localization",
-    "internationalization",
-    "fluent"
-  ],
-  "license": "MIT",
-  "main": "lib/index.js",
-  "oclif": {
-    "commands": "./lib/commands",
-    "bin": "inlang",
-    "plugins": [
-      "@oclif/plugin-help",
-      "@oclif/plugin-not-found"
-    ]
-  },
-  "repository": "inlang/inlang",
-  "scripts": {
-    "postpack": "rm -f oclif.manifest.json",
-    "posttest": "eslint . --ext .ts --config .eslintrc",
-    "prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
-    "version": "oclif-dev readme && git add README.md",
-    "format": "prettier .",
-    "format:fix": "npm run format -- --write",
-    "lint": "eslint .",
-    "lint:fix": "npm run lint -- --fix",
-    "test": "jest",
-    "test:watch": "npm test -- --watch"
-  },
-  "types": "lib/index.d.ts"
-}
