--- apps/cli-old/test/hello.test.ts
@@ -1,4 +1,0 @@
-it('shoud be Hello World', () => {
-  const x = 'Hello World';
-  expect(x).toEqual('Hello World');
-});
