--- apps/cli-old/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "importHelpers": true,
-    "module": "commonjs",
-    "outDir": "lib",
-    "rootDir": "src",
-    "strict": true,
-    "target": "es2017"
-  },
-  "include": [
-    "src/**/*"
-  ]
-}
