+++ apps/cli/.editorconfig
@@ -1,11 +1,17 @@
+; top-most EditorConfig file
 root = true
 
+; Unix-style newlines
 [*]
-indent_style = space
-indent_size = 2
 charset = utf-8
-trim_trailing_whitespace = true
+end_of_line = lf
 insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.ts]
+indent_style = space
+indent_size = 4
+max_line_length = 120
 
 [*.md]
-trim_trailing_whitespace = false
+max_line_length = 80
