+++ apps/cli/.eslintignore
@@ -1,1 +1,7 @@
-/dist
+# /node_modules/* in the project root is ignored by default
+# build artefacts
+dist/*
+jest.config.js
+tests/*
+jest.config.js
+bin/*
