+++ apps/cli/.eslintrc
@@ -1,3 +1,9 @@
 {
+  "root": true,
+  "env": {
+    "browser": false,
+    "node": true
+  },
+  "parser": "@typescript-eslint/parser",
   "extends": ["@inlang/eslint-config"]
 }
