--- apps/cli/.mocharc.json
@@ -1,12 +1,0 @@
-{
-  "require": [
-    "test/helpers/init.js",
-    "ts-node/register"
-  ],
-  "watch-extensions": [
-    "ts"
-  ],
-  "recursive": true,
-  "reporter": "spec",
-  "timeout": 60000
-}
