+++ apps/cli/.prettierignore
