+++ apps/cli/.prettierrc
