--- apps/cli/bin/dev.cmd
@@ -1,3 +1,0 @@
-@echo off
-
-node "%~dp0\dev" %*
