--- apps/cli/bin/run
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-
-import * as oclif from "@oclif/core";
-
-oclif
-  .run()
-  .then(require("@oclif/core/flush"))
-  .catch(require("@oclif/core/handle"));
