+++ apps/cli/package.json
@@ -1,77 +1,36 @@
 {
   "name": "@inlang/cli",
-  "version": "0.0.0",
-  "description": "TODO",
-  "author": "inlang",
+  "version": "0.1.0",
+  "private": true,
+  "main": "./dist/main.js",
+  "types": "./dist/main.d.ts",
   "type": "module",
   "bin": {
-    "inlang": "./bin/run"
+    "inlang": "./bin/run.js"
   },
-  "homepage": "https://github.com/inlang/inlang",
-  "license": "Apache 2.0",
-  "main": "dist/index.js",
-  "repository": "inlang/inlang",
-  "files": [
-    "/bin",
-    "/dist",
-    "/npm-shrinkwrap.json",
-    "/oclif.manifest.json"
-  ],
-  "dependencies": {
-    "@oclif/core": "^1",
-    "@oclif/plugin-help": "^5",
-    "@oclif/plugin-plugins": "^2.0.1",
-    "node-fetch": "^2.6.6"
+  "scripts": {
+    "build": "rollup --config",
+    "dev": "rollup --config --watch",
+    "clean": "rm -r dist",
+    "format": "prettier .",
+    "format:fix": "npm run format -- --write",
+    "lint": "eslint .",
+    "lint:fix": "npm run lint -- --fix",
+    "test": "jest --passWithNoTests",
+    "test:watch": "npm test -- --watch"
   },
   "devDependencies": {
     "@inlang/eslint-config": "*",
-    "@oclif/test": "^2",
-    "@types/chai": "^4",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.9.4",
-    "chai": "^4",
-    "eslint": "^7.32.0",
-    "globby": "^11",
-    "mocha": "^9",
-    "oclif": "^2",
-    "shx": "^0.3.3",
-    "ts-node": "^10.2.1",
-    "tslib": "^2.3.1",
-    "typescript": "^4.4.3"
+    "@rollup/plugin-commonjs": "^21.0.1",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-typescript": "^8.3.0",
+    "rollup": "^2.66.0"
   },
-  "oclif": {
-    "bin": "inlang",
-    "dirname": "oex",
-    "commands": "./dist/commands",
-    "plugins": [
-      "@oclif/plugin-help",
-      "@oclif/plugin-plugins"
-    ]
-  },
-  "scripts": {
-    "build": "shx rm -rf dist && tsc -b",
-    "lint": "eslint . --ext .ts --config .eslintrc",
-    "postpack": "shx rm -f oclif.manifest.json",
-    "posttest": "yarn lint",
-    "prepack": "yarn build && oclif manifest && oclif readme",
-    "test": "mocha --forbid-only \"test/**/*.test.ts\"",
-    "version": "oclif readme && git add README.md"
-  },
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "bugs": "https://github.com/inlang/inlang/issues",
-  "keywords": [
-    "inlang",
-    "cli",
-    "i18n",
-    "internationalization",
-    "internationalize",
-    "localize",
-    "lokalise",
-    "localization",
-    "translate",
-    "translation"
-  ],
-  "types": "dist/index.d.ts"
+  "dependencies": {
+    "@inlang/adapters": "*",
+    "@inlang/common": "*",
+    "cleye": "^1.0.1",
+    "consola": "^2.15.3",
+    "node-fetch": "^3.2.0"
+  }
 }
