+++ apps/cli/rollup.config.js
@@ -1,11 +1,24 @@
-import typescript from "@rollup/plugin-typescript";
-import { nodeResolve } from "@rollup/plugin-node-resolve";
+import typescript from '@rollup/plugin-typescript';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import commonjs from '@rollup/plugin-commonjs';
+
+/**
+ * What is rollup used for?
+ *
+ * Rollup bundles all dependencies into one file i.e.
+ * the resulting code has no dependencies on node modules.
+ * That avoids problems that can (did) occur when resolving
+ * those dependencies.
+ */
 
 export default {
-  input: "src/index.ts",
+  input: 'src/main.ts',
   output: {
-    dir: "dist",
-    format: "commonjs",
+    dir: 'dist',
+    format: 'es',
   },
-  plugins: [nodeResolve(), typescript()],
+  // nodeResolve = bundle the dependencies
+  // typescript = compile typescript
+  // commonjs = because of commonjs peer dependencies (peggy.js)
+  plugins: [nodeResolve(), typescript(), commonjs()],
 };
