+++ apps/cli/src/commands/upload.ts
