+++ apps/cli/src/main.ts
