--- apps/cli/test/helpers/init.js
@@ -1,6 +1,0 @@
-const path = require('path')
-process.env.TS_NODE_PROJECT = path.resolve('test/tsconfig.json')
-process.env.NODE_ENV = 'development'
-
-global.oclif = global.oclif || {}
-global.oclif.columns = 80
