+++ apps/cli/tsconfig.json
@@ -1,13 +1,16 @@
 {
   "compilerOptions": {
-    "declaration": true,
-    "importHelpers": true,
-    "skipLibCheck": true,
-    "moduleResolution": "node",
-    "outDir": "dist",
-    "rootDir": "src",
+    "esModuleInterop": true,
     "strict": true,
-    "target": "es2019"
+    "target": "ES2020",
+    "module": "ES2020",
+    "noImplicitAny": true,
+    "moduleResolution": "node",
+    "skipLibCheck": true,
+    "outDir": "dist",
+    "removeComments": true,
+    "declaration": false,
+    "allowJs": false
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/**.ts"]
 }
