--- apps/cli3/.editorconfig
@@ -1,17 +1,0 @@
-; top-most EditorConfig file
-root = true
-
-; Unix-style newlines
-[*]
-charset = utf-8
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.ts]
-indent_style = space
-indent_size = 4
-max_line_length = 120
-
-[*.md]
-max_line_length = 80
