--- apps/cli3/.eslintignore
@@ -1,7 +1,0 @@
-# /node_modules/* in the project root is ignored by default
-# build artefacts
-dist/*
-jest.config.js
-tests/*
-jest.config.js
-bin/*
