--- apps/cli3/.eslintrc
@@ -1,9 +1,0 @@
-{
-  "root": true,
-  "env": {
-    "browser": false,
-    "node": true
-  },
-  "parser": "@typescript-eslint/parser",
-  "extends": ["@inlang/eslint-config"]
-}
