--- apps/cli3/.prettierignore
@@ -1,4 +1,0 @@
-# /node_modules/* in the project root is ignored by default
-# build artefacts
-dist
-package-lock.json
