--- apps/cli3/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "printWidth": 120,
-  "singleQuote": true
-}
