--- apps/cli3/README.md
@@ -1,3 +1,0 @@
-# @inlang/cli
-
-Intended to use Oclif but oclif does not support esmodules (yikes).
