--- apps/cli3/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@inlang/cli3",
-  "version": "0.1.0",
-  "private": true,
-  "main": "./dist/main.js",
-  "types": "./dist/main.d.ts",
-  "type": "module",
-  "bin": {
-    "inlang": "./bin/run.js"
-  },
-  "scripts": {
-    "build": "rollup --config",
-    "dev": "rollup --config --watch",
-    "clean": "rm -r dist",
-    "format": "prettier .",
-    "format:fix": "npm run format -- --write",
-    "lint": "eslint .",
-    "lint:fix": "npm run lint -- --fix",
-    "test": "jest --passWithNoTests",
-    "test:watch": "npm test -- --watch"
-  },
-  "devDependencies": {
-    "@inlang/eslint-config": "*",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.1.3",
-    "@rollup/plugin-typescript": "^8.3.0",
-    "rollup": "^2.66.0"
-  },
-  "dependencies": {
-    "@inlang/adapters": "*",
-    "@inlang/common": "*",
-    "cleye": "^1.0.1",
-    "consola": "^2.15.3",
-    "node-fetch": "^3.2.0"
-  }
-}
