--- apps/cli3/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "strict": true,
-    "target": "ES2020",
-    "module": "ES2020",
-    "noImplicitAny": true,
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "outDir": "dist",
-    "removeComments": true,
-    "declaration": false,
-    "allowJs": false
-  },
-  "include": ["src/**/**.ts"]
-}
